pyproject.toml
[tool.poetry]
name = "autocaption"
version = "0.1.0"
description = "Automatically transcribe audio/visual files to a variety of subtitle formats."
authors = ["Daniel Maxwell-Ross <daniel@maxwell-ross.us>"]
license = "GNU GPL v3"

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-speech = "^2.0.1" # https://googleapis.dev/python/speech/2.0.1/index.html
google-cloud-storage = "^1.35.0" # https://googleapis.dev/python/storage/1.35.0/index.html
click = "^7.1.2"
pysubs2 = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
autocaption = 'autocaption.main:main'